Admin and user functionality are all done for first version. Also made schema specifications work.

+++ api/bp_user/backend.py
@@ -1,31 +1,25 @@
+from flask import g
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm.exc import NoResultFound
 
 from ..common.exceptions import RecordAlreadyExists, RecordNotFound, MissingArguments
+from ..common.exceptions import CannotChangeOthersProfile, CannotDeleteOthersProfile
 
 from ..common.models import User
-from ..common.database import db_session
+
+
 
 def create_user(user_data):
-
     if user_data['email'] is None or user_data['password'] is None: 
         msg = "Please provide an email and a password."
         raise MissingArguments(message=msg)
-
-    
     user = User(**user_data)
-    print("backend.the user data is: ", user_data)
-    print("backend.The created user is: ", user)
     user.set_password(user_data['password'])
-    print("backend.The user is: ", user)
     try:
         user.save()
-        print('backend.user saved')
-        # db_session.commit()
     except IntegrityError:
         msg = 'Email `%s` is already in use for another account.' % user_data['email']
         raise RecordAlreadyExists(message=msg)
-
     return user
 
 
@@ -35,7 +29,6 @@ def get_user_by_id(user_id):
     except NoResultFound:
         msg = 'There is no User with `id: %s`' % id
         raise RecordNotFound(message=msg)
-
     return result
 
 
@@ -45,11 +38,18 @@ def get_all_users():
 
 def update_user(user_data, user_id):
     user = get_user_by_id(user_id)
-    user.update(**user_data)
-
-    return user
+    if user.email == g.current_user.email:
+        user.update(**user_data)
+        return user
+    else:
+        msg = "You can't change other people's profile."
+        raise CannotChangeOthersProfile(message=msg)
 
 
 def delete_user(user_id):
     user = get_user_by_id(user_id)
-    user.delete()
+    if user.email == g.current_user.email:
+        user.delete()
+    else:
+        msg = "You can't delete other people's profile."
+        raise CannotDeleteOthersProfile(message=msg)
